pyproject.toml
[project]
name = "makoto-lite-llm"
version = "0.1.0"
description = "LLM proxy gateway with auth and multi-tenant management"
readme = "README.md"
authors = [
    { name = "Makoto Sandiga", email = "makoto.sandiga@nubank.com.br" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cachetools>=7.0.5",
    "cryptography>=46.0.6",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-multipart>=0.0.22",
    "redis[hiredis]>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uuid7>=0.1.0",
    "uvicorn[standard]>=0.42.0",
]

[project.optional-dependencies]
e2e = [
    "playwright>=1.48.0",
    "pytest-playwright>=0.6.2",
    "testcontainers[keycloak]>=4.14.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
    "ruff>=0.15.8",
    "testcontainers[postgres,redis]>=4.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-m 'not e2e and not live'"
markers = [
    "e2e: full end-to-end tests that boot containers and a browser; run with `pytest -m e2e`",
    "live: tests that hit real third-party APIs and need credentials; run with `pytest -m live`",
]
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "SIM"]


